Move customer info creation into CheckoutSessionLoader and ElementsSessionLoader
Each loader now fully owns its path end-to-end: fetching the session
and creating the appropriate CustomerInfo variant. SessionResult returns
customerInfo directly instead of leaking CheckoutSessionResponse to the
caller. CustomerInfo and toCustomerInfo() move to LoadSession.kt as
top-level types shared by both loaders and the loader.

Committed-By-Agent: claude
