+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/state/ElementsSessionLoaderTest.kt
@@ -6,6 +6,7 @@ import com.stripe.android.model.PaymentIntentFixtures
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PaymentSheetFixtures
 import com.stripe.android.paymentsheet.model.SavedSelection
+import com.stripe.android.testing.FakeErrorReporter
 import com.stripe.android.utils.FakeElementsSessionRepository
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
@@ -24,7 +25,7 @@ internal class ElementsSessionLoaderTest {
 
         assertThat(result.elementsSession.stripeIntent)
             .isEqualTo(PaymentIntentFixtures.PI_REQUIRES_PAYMENT_METHOD)
-        assertThat(result.checkoutSession).isNull()
+        assertThat(result.customerInfo).isNull()
     }
 
     @Test
@@ -58,6 +59,7 @@ internal class ElementsSessionLoaderTest {
         )
         val loader = ElementsSessionLoader(
             elementsSessionRepository = elementsSessionRepository,
+            errorReporter = FakeErrorReporter(),
         )
         Scenario(
             loader = loader,
